requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cloudscraper==1.2.58
colorama==0.4.4
cryptography==36.0.1
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-import-export==2.8.0
et-xmlfile==1.1.0
fonttools==4.32.0
html5lib==1.1
idna==3.3
iso8601==1.0.2
m3u8==1.0.0
MarkupPy==1.14
mysql-connector-python==8.0.28
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.0
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyphen==0.12.0
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tinycss2==1.1.1
toml==0.10.2
tzdata==2021.5
Unidecode==1.3.2
urllib3==1.26.8
weasyprint==54.3
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1
